requirements.txt
# Core ML and Vision
torch>=2.0.0
transformers>=4.53.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0

# Text-to-Speech
pyttsx3>=2.90

# Web interface (optional)
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
loguru>=0.7.0

# For Gemma models
accelerate>=0.24.0
bitsandbytes>=0.41.0
